+++ boards/adafruit-feather-rp2040/examples/adafruit_feather_neopixel_rainbow.rs
@@ -20,7 +20,7 @@ use adafruit_feather_rp2040::{
 };
 use core::iter::once;
 use embedded_hal::timer::CountDown;
-use embedded_time::duration::Extensions;
+use fugit::ExtU32;
 use panic_halt as _;
 use smart_leds::{brightness, SmartLedsWrite, RGB8};
 use ws2812_pio::Ws2812;
@@ -71,7 +71,7 @@ fn main() -> ! {
         ws.write(brightness(once(wheel(n)), 32)).unwrap();
         n = n.wrapping_add(1);
 
-        delay.start(25.milliseconds());
+        delay.start(25.millis());
         let _ = nb::block!(delay.wait());
     }
 }