+++ boards/adafruit-macropad/examples/adafruit-macropad_blinky.rs
@@ -14,7 +14,6 @@ use adafruit_macropad::{
     Pins, XOSC_CRYSTAL_FREQ,
 };
 use embedded_hal::digital::v2::OutputPin;
-use embedded_time::rate::*;
 use panic_halt as _;
 
 /// Entry point to our bare-metal application.
@@ -40,7 +39,7 @@ fn main() -> ! {
     .ok()
     .unwrap();
 
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     let sio = Sio::new(pac.SIO);
     let pins = Pins::new(