+++ boards/rp-pico/examples/pico_uart_irq_buffer.rs
@@ -21,9 +21,6 @@
 // objects as generic embedded devices.
 use embedded_hal::{digital::v2::OutputPin, serial::Write as UartWrite};
 
-// We need this for the 'Delay' object to work.
-use embedded_time::fixed_point::FixedPoint;
-
 // The writeln! trait.
 use core::fmt::Write;
 
@@ -114,7 +111,7 @@ fn main() -> ! {
     .unwrap();
 
     // Lets us wait for fixed periods of time
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     // The single-cycle I/O block controls our GPIO pins
     let sio = hal::Sio::new(pac.SIO);