+++ rp2040-hal/src/uart/common_configs.rs
@@ -1,9 +1,10 @@
+use fugit::HertzU32;
+
 use super::{DataBits, StopBits, UartConfig};
-use embedded_time::rate::Baud;
 
 /// 9600 baud, 8 data bits, no parity, 1 stop bit
 pub const _9600_8_N_1: UartConfig = UartConfig {
-    baudrate: Baud(9600),
+    baudrate: HertzU32::from_raw(9600),
     data_bits: DataBits::Eight,
     stop_bits: StopBits::One,
     parity: None,
@@ -11,7 +12,7 @@ pub const _9600_8_N_1: UartConfig = UartConfig {
 
 /// 19200 baud, 8 data bits, no parity, 1 stop bit
 pub const _19200_8_N_1: UartConfig = UartConfig {
-    baudrate: Baud(19200),
+    baudrate: HertzU32::from_raw(19200),
     data_bits: DataBits::Eight,
     stop_bits: StopBits::One,
     parity: None,
@@ -19,7 +20,7 @@ pub const _19200_8_N_1: UartConfig = UartConfig {
 
 /// 38400 baud, 8 data bits, no parity, 1 stop bit
 pub const _38400_8_N_1: UartConfig = UartConfig {
-    baudrate: Baud(38400),
+    baudrate: HertzU32::from_raw(38400),
     data_bits: DataBits::Eight,
     stop_bits: StopBits::One,
     parity: None,
@@ -27,7 +28,7 @@ pub const _38400_8_N_1: UartConfig = UartConfig {
 
 /// 57600 baud, 8 data bits, no parity, 1 stop bit
 pub const _57600_8_N_1: UartConfig = UartConfig {
-    baudrate: Baud(57600),
+    baudrate: HertzU32::from_raw(57600),
     data_bits: DataBits::Eight,
     stop_bits: StopBits::One,
     parity: None,
@@ -35,7 +36,7 @@ pub const _57600_8_N_1: UartConfig = UartConfig {
 
 /// 115200 baud, 8 data bits, no parity, 1 stop bit
 pub const _115200_8_N_1: UartConfig = UartConfig {
-    baudrate: Baud(115200),
+    baudrate: HertzU32::from_raw(115200),
     data_bits: DataBits::Eight,
     stop_bits: StopBits::One,
     parity: None,