+++ boards/adafruit-macropad/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use `rp2040-hal`'s entry function.
+- Migrate from `embedded-time` to `fugit`
 
 ## 0.3.0 - 2022-06-13