+++ boards/seeeduino-xiao-rp2040/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- None
+- Migrate from `embedded-time` to `fugit`
 
 ## 0.1.0 - 2022-06-26